requirements.txt
torch
tqdm
requests
pyyaml
pydicom
numpy